Refactor | Extract additional generator files
