+++ src/gen-xml.ts
@@ -40,7 +40,8 @@ import {
 	inch2Emu,
 	valToPts,
 } from './gen-utils';
-import * as bullets from './generators/bullets';
+import * as paragraph from './generators/paragraph';
+import textRun from './generators/textRun';
 
 const ImageSizingXml = {
 	cover: function (imgSize: { w: number; h: number }, boxDim: { w: number; h: number; x: number; y: number }) {
@@ -838,167 +839,6 @@ export function slideObjectRelationsToXml (slide: PresSlide | SlideLayout, defau
 }
 
 /**
- * Generate XML Paragraph Properties
- * @param {ISlideObject|TextProps} textObj - text object
- * @param {boolean} isDefault - array of default relations
- * @param slide
- * @return {string} XML
- */
-export function genXmlParagraphProperties (textObj: ISlideObject | TextProps, isDefault: boolean, slide: PresSlide | SlideLayout): string {
-	let strXmlBullet = '';
-	let strXmlLnSpc = '';
-	let strXmlParaSpc = '';
-	let strXmlTabStops = '';
-	const tag = isDefault ? 'a:lvl1pPr' : 'a:pPr';
-
-	let paragraphPropXml = `<${tag}${textObj.options.rtlMode ? ' rtl="1" ' : ''}`;
-
-	// A: Build paragraphProperties
-	{
-		// OPTION: align
-		if (textObj.options.align) {
-			switch (textObj.options.align) {
-				case 'left':
-					paragraphPropXml += ' algn="l"';
-					break;
-				case 'right':
-					paragraphPropXml += ' algn="r"';
-					break;
-				case 'center':
-					paragraphPropXml += ' algn="ctr"';
-					break;
-				case 'justify':
-					paragraphPropXml += ' algn="just"';
-					break;
-				default:
-					paragraphPropXml += '';
-					break;
-			}
-		}
-
-		if (textObj.options.lineSpacing) {
-			strXmlLnSpc = `<a:lnSpc><a:spcPts val="${Math.round(textObj.options.lineSpacing * 100)}"/></a:lnSpc>`;
-		} else if (textObj.options.lineSpacingMultiple) {
-			strXmlLnSpc = `<a:lnSpc><a:spcPct val="${Math.round(textObj.options.lineSpacingMultiple * 100000)}"/></a:lnSpc>`;
-		}
-
-		// OPTION: indent
-		if (textObj.options.indentLevel && !isNaN(Number(textObj.options.indentLevel)) && textObj.options.indentLevel > 0) {
-			paragraphPropXml += ` lvl="${textObj.options.indentLevel}"`;
-		}
-
-		// OPTION: Paragraph Spacing: Before/After
-		if (textObj.options.paraSpaceBefore && !isNaN(Number(textObj.options.paraSpaceBefore)) && textObj.options.paraSpaceBefore > 0) {
-			strXmlParaSpc += `<a:spcBef><a:spcPts val="${Math.round(textObj.options.paraSpaceBefore * 100)}"/></a:spcBef>`;
-		}
-		if (textObj.options.paraSpaceAfter && !isNaN(Number(textObj.options.paraSpaceAfter)) && textObj.options.paraSpaceAfter > 0) {
-			strXmlParaSpc += `<a:spcAft><a:spcPts val="${Math.round(textObj.options.paraSpaceAfter * 100)}"/></a:spcAft>`;
-		}
-
-		// OPTION: bullet
-		if (textObj.options.bullet) {
-			// The slide will be mutated and a new image rel added when the bullet includes a new icon
-			//
-			const imageRid = bullets.maybeAddImageRel(textObj.options, (slide as PresSlide));
-			const bulletXml = bullets.generateXml(textObj.options, slide, imageRid);
-			paragraphPropXml += bullets.paragraphPropXml(textObj.options);
-			strXmlBullet = bulletXml;
-		}
-
-		// OPTION: tabStops
-		if (textObj.options.tabStops && Array.isArray(textObj.options.tabStops)) {
-			const tabStopsXml = textObj.options.tabStops.map(stop => `<a:tab pos="${inch2Emu(stop.position || 1)}" algn="${stop.alignment || 'l'}"/>`).join('');
-			strXmlTabStops = `<a:tabLst>${tabStopsXml}</a:tabLst>`;
-		}
-
-		// B: Close Paragraph-Properties
-		// IMPORTANT: strXmlLnSpc, strXmlParaSpc, and strXmlBullet require strict ordering - anything out of order is ignored. (PPT-Online, PPT for Mac)
-		paragraphPropXml += '>' + strXmlLnSpc + strXmlParaSpc + strXmlBullet + strXmlTabStops;
-		if (isDefault) paragraphPropXml += genXmlTextRunProperties(textObj.options, true);
-		paragraphPropXml += '</' + tag + '>';
-	}
-
-	return paragraphPropXml;
-}
-
-/**
- * Generate XML Text Run Properties (`a:rPr`)
- * @param {ObjectOptions|TextPropsOptions} opts - text options
- * @param {boolean} isDefault - whether these are the default text run properties
- * @return {string} XML
- */
-export function genXmlTextRunProperties (opts: ObjectOptions | TextPropsOptions, isDefault: boolean): string {
-	let runProps = '';
-	const runPropsTag = isDefault ? 'a:defRPr' : 'a:rPr';
-
-	// BEGIN runProperties (ex: `<a:rPr lang="en-US" sz="1600" b="1" dirty="0">`)
-	runProps += '<' + runPropsTag + ' lang="' + (opts.lang ? opts.lang : 'en-US') + '"' + (opts.lang ? ' altLang="en-US"' : '');
-	runProps += opts.fontSize ? ` sz="${Math.round(opts.fontSize * 100)}"` : ''; // NOTE: Use round so sizes like '7.5' wont cause corrupt presentations
-	runProps += opts?.bold ? ` b="${opts.bold ? '1' : '0'}"` : '';
-	runProps += opts?.italic ? ` i="${opts.italic ? '1' : '0'}"` : '';
-
-	runProps += opts?.strike ? ` strike="${typeof opts.strike === 'string' ? opts.strike : 'sngStrike'}"` : '';
-	if (typeof opts.underline === 'object' && opts.underline?.style) {
-		runProps += ` u="${opts.underline.style}"`;
-	} else if (typeof opts.underline === 'string') {
-		// DEPRECATED: opts.underline is an object as of v3.5.0
-		runProps += ` u="${String(opts.underline)}"`;
-	} else if (opts.hyperlink) {
-		runProps += ' u="sng"';
-	}
-	if (opts.baseline) {
-		runProps += ` baseline="${Math.round(opts.baseline * 50)}"`;
-	} else if (opts.subscript) {
-		runProps += ' baseline="-40000"';
-	} else if (opts.superscript) {
-		runProps += ' baseline="30000"';
-	}
-	runProps += opts.charSpacing ? ` spc="${Math.round(opts.charSpacing * 100)}" kern="0"` : ''; // IMPORTANT: Also disable kerning; otherwise text won't actually expand
-	runProps += ' dirty="0">';
-	// Color / Font / Highlight / Outline are children of <a:rPr>, so add them now before closing the runProperties tag
-	if (opts.color || opts.fontFace || opts.outline || (typeof opts.underline === 'object' && opts.underline.color)) {
-		if (opts.outline && typeof opts.outline === 'object') {
-			runProps += `<a:ln w="${valToPts(opts.outline.size || 0.75)}">${genXmlColorSelection(opts.outline.color || 'FFFFFF')}</a:ln>`;
-		}
-		if (opts.color) runProps += genXmlColorSelection({ color: opts.color, transparency: opts.transparency });
-		if (opts.highlight) runProps += `<a:highlight>${createColorElement(opts.highlight)}</a:highlight>`;
-		if (typeof opts.underline === 'object' && opts.underline.color) runProps += `<a:uFill>${genXmlColorSelection(opts.underline.color)}</a:uFill>`;
-		if (opts.glow) runProps += `<a:effectLst>${createGlowElement(opts.glow, DEF_TEXT_GLOW)}</a:effectLst>`;
-		if (opts.fontFace) {
-			// NOTE: 'cs' = Complex Script, 'ea' = East Asian (use "-120" instead of "0" - per Issue #174); ea must come first (Issue #174)
-			runProps += `<a:latin typeface="${opts.fontFace}" pitchFamily="34" charset="0"/><a:ea typeface="${opts.fontFace}" pitchFamily="34" charset="-122"/><a:cs typeface="${opts.fontFace}" pitchFamily="34" charset="-120"/>`;
-		}
-	}
-
-	// Hyperlink support
-	if (opts.hyperlink) {
-		if (typeof opts.hyperlink !== 'object') throw new Error('ERROR: text `hyperlink` option should be an object. Ex: `hyperlink:{url:\'https://github.com\'}` ');
-		else if (!opts.hyperlink.url && !opts.hyperlink.slide) throw new Error('ERROR: \'hyperlink requires either `url` or `slide`\'');
-		else if (opts.hyperlink.url) {
-			// runProps += '<a:uFill>'+ genXmlColorSelection('0000FF') +'</a:uFill>'; // Breaks PPT2010! (Issue#74)
-			runProps += `<a:hlinkClick r:id="rId${opts.hyperlink._rId}" invalidUrl="" action="" tgtFrame="" tooltip="${opts.hyperlink.tooltip ? encodeXmlEntities(opts.hyperlink.tooltip) : ''
-			}" history="1" highlightClick="0" endSnd="0"${opts.color ? '>' : '/>'}`;
-		} else if (opts.hyperlink.slide) {
-			runProps += `<a:hlinkClick r:id="rId${opts.hyperlink._rId}" action="ppaction://hlinksldjump" tooltip="${opts.hyperlink.tooltip ? encodeXmlEntities(opts.hyperlink.tooltip) : ''
-			}"${opts.color ? '>' : '/>'}`;
-		}
-		if (opts.color) {
-			runProps += ' <a:extLst>';
-			runProps += '  <a:ext uri="{A12FA001-AC4F-418D-AE19-62706E023703}">';
-			runProps += '   <ahyp:hlinkClr xmlns:ahyp="http://schemas.microsoft.com/office/drawing/2018/hyperlinkcolor" val="tx"/>';
-			runProps += '  </a:ext>';
-			runProps += ' </a:extLst>';
-			runProps += '</a:hlinkClick>';
-		}
-	}
-
-	// END runProperties
-	runProps += `</${runPropsTag}>`;
-
-	return runProps;
-}
-
-/**
  * Build textBody text runs [`<a:r></a:r>`] for paragraphs [`<a:p>`]
  * @param {TextProps} textObj - Text object
  * @return {string} XML string
@@ -1007,7 +847,7 @@ export function genXmlTextRun (textObj: TextProps): string {
 	// If text string has line-breaks, split sequential runs with a `<a:br/>` tag for separation
 	const textRuns = [];
 	let xmlTextRun = '';
-	const xmlProperties = genXmlTextRunProperties(textObj.options, false);
+	const xmlProperties = textRun.properties.generateXml(textObj.options, false);
 	if (textObj.text && typeof textObj.text === 'string') {
 		textObj.text.split(CRLF).forEach(line => {
 			textRuns.push(line);
@@ -1134,7 +974,7 @@ export function genXmlTextBody (slideObj: ISlideObject | TableCell, slide: PresS
 		// NOTE: shape type 'LINE' has different text align needs (a lstStyle.lvl1pPr between bodyPr and p)
 		// FIXME: LINE horiz-align doesnt work (text is always to the left inside line) (FYI: the PPT code diff is substantial!)
 		if (opts.h === 0 && opts.line && opts.align) strSlideXml += '<a:lstStyle><a:lvl1pPr algn="l"/></a:lstStyle>';
-		else if (slideObj._type === 'placeholder') strSlideXml += `<a:lstStyle>${genXmlParagraphProperties(slideObj, true, slide)}</a:lstStyle>`;
+		else if (slideObj._type === 'placeholder') strSlideXml += `<a:lstStyle>${paragraph.generateXml(slideObj, true, slide)}</a:lstStyle>`;
 		else strSlideXml += '<a:lstStyle/>';
 	}
 
@@ -1236,7 +1076,7 @@ export function genXmlTextBody (slideObj: ISlideObject | TableCell, slide: PresS
 			textObj.options.indentLevel = textObj.options.indentLevel || opts.indentLevel;
 			textObj.options.paraSpaceBefore = textObj.options.paraSpaceBefore || opts.paraSpaceBefore;
 			textObj.options.paraSpaceAfter = textObj.options.paraSpaceAfter || opts.paraSpaceAfter;
-			paragraphPropXml = genXmlParagraphProperties(textObj, false, slide);
+			paragraphPropXml = paragraph.generateXml(textObj, false, slide);
 
 			strSlideXml += paragraphPropXml.replace('<a:pPr></a:pPr>', ''); // IMPORTANT: Empty "pPr" blocks will generate needs-repair/corrupt msg
 			// C: Inherit any main options (color, fontSize, etc.)
