+++ src/generators/bullet.ts
