+++ src/generators/textRun/index.ts
@@ -1,0 +1,5 @@
+import * as properties from './properties';
+
+export default {
+	properties,
+};
